Move IBoreholePermissionService
