+++ src/api/Controllers/BoreholeControllerBase.cs
@@ -1,4 +1,5 @@
 ﻿using BDMS.Models;
+using BDMS.Services;
 using Microsoft.AspNetCore.Mvc;
 
 namespace BDMS.Controllers;
