+++ src/api/Controllers/BoreholeGeometryController.cs
@@ -1,6 +1,7 @@
 ﻿using BDMS.Authentication;
 using BDMS.BoreholeGeometry;
 using BDMS.Models;
+using BDMS.Services;
 using CsvHelper;
 using CsvHelper.TypeConversion;
 using Microsoft.AspNetCore.Authorization;
