+++ src/api/Controllers/LithologicalDescriptionController.cs
@@ -1,6 +1,7 @@
 ﻿using Amazon;
 using BDMS.Authentication;
 using BDMS.Models;
+using BDMS.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
