+++ tests/api/Controllers/BoreholeGeometryControllerTest.cs
@@ -1,4 +1,5 @@
 ﻿using BDMS.Models;
+using BDMS.Services;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
