+++ tests/api/Controllers/DocumentControllerTest.cs
@@ -1,4 +1,5 @@
 ﻿using BDMS.Models;
+using BDMS.Services;
 using Microsoft.AspNetCore.Http;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
